requirements.txt
Flask_Bcrypt==0.7.1
Flask_SQLAlchemy==2.4.1
Flask==1.1.1
Flask_Migrate==2.5.2
Flask_Script==2.0.6
SQLAlchemy==1.3.10
Flask_RESTful==0.3.7
alembic==1.3.0
PyJWT==1.7.1
